requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blessed==1.20.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
contourpy==1.1.1
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
dnspython==2.6.1
docker-pycreds==0.4.0
email_validator==2.1.1
exceptiongroup==1.2.1
face-alignment==1.3.5
facenet-pytorch==2.5.2
fast-bss-eval==0.1.4
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
gpustat==1.1.1
gradio==4.33.0
gradio_client==0.17.0
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.2
llvmlite==0.41.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmcv==1.7.1
mmengine==0.10.4
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numba==0.58.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.7.0.68
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.0
ordered-set==4.1.0
orjson==3.10.3
oss2==2.17.0
packaging==24.0
pandas==1.5.1
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
protobuf==3.20.1
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==1.6.4
pytorch-ranger==0.1.1
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.4.2
rpds-py==0.18.1
rsa==4.9
ruff==0.4.7
scikit-image==0.21.0
scipy==1.10.1
semantic-version==2.10.0
sentry-sdk==2.3.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
tifffile==2023.7.10
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==1.13.0+cu116
torch-mir-eval==0.4
torch-optimizer==0.3.0
torchaudio==0.13.0+cu116
torchmetrics==1.4.0.post0
torchvision==0.14.0+cu116
tqdm==4.65.2
triton==2.2.0
typer==0.12.3
typing_extensions==4.12.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
wandb==0.17.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
yapf==0.40.2
yarl==1.9.4
zipp==3.19.1
